applets/manage-inputmethod: still show close button out of Touch Mode
The virtual keyboard can appear with a touch input, not just when
you're in Touch Mode. However b1888b5e9ecf864935a55c0878b1aa8578bfe5ca
made the button only appear when in Touch Mode, giving the user no
clear way to close the virtual keyboard when it was invoked outside of
Touch Mode.

This commit fixes that by always showing the close button when the
keyboard is visible due to being activated with a touch event outside of
Touch mode. There is no change for the "disable virtual
keyboard" button.