kcms/nightcolor: Make Night Color enabled/disabled status part of the combobox
Currently we have a situation where in order to change the Night Color mode,
you have to go through Automatic mode first (and thereby make a geolocation
network request) as it's the default setting. With this commit, you turn on
Night Color and choose an activation mode all in a single step using the
combobox, bypassing this issue and simplifying the UI too.