Add KDED module for adjusting the system time zone based on location
We don't need accurate location data, so it just talks to the KDE geoip endpoint
directly and adjusts the timezone accordingly.

It does that only on a network with full connectivity (ie. not behind a portal),
that is not metered and not connected to a VPN (to avoid using the VPN gateway's
timezone inadvertendly) and only a maximum of once per hour.

The module is off-by-default to not do unsolicited network requests
and a UI to enable it (in date time KCM and/or digital clock applet settings)
is tbd.