Drop direct session switching from krunner
KRunner used to be the sole location for session switching back in early
Plasma 5. This was mostly consolidated into the screenlocker which makes
life secure and easy as we can ensure the screen is locked before
switching, something this codepath fails to do.

Searching for "Switch User" still works and opens the relevant
lockscreen page, so it's only one additional click away.