kcms/notifications: Exit System Notifications page on invalid source
When the model resets (by pressing the Defaults or Reset buttons),
the app notification gets also reset and de-assigned from any
application (invalid rootIndex).

For the System Notifications page, with no application list, this
just shows a misleading placeholder text, because we cannot select
any app here. So let's go back to the main page.

It would be better not to reset the full model, or not to deselect
the app, but for now let's avoid presenting a "broken" UI.