startkde: add new libexec tools to configure distro live sessions
this is a port of scripts from fedora and neon that mangle the live
session so it doesn't fall apart due to baloo or lockscreens. I hope
it's of use for others too

due to the nature of the beast this is fairly hard to test in practice
so I've added some auto tests instead but I do fully expect some growing
pains initially.

please note the README for how to use the two programs